Added simple build configuration file

+++ PicoDotNet.Builder.Runtime/Program.cs
@@ -8,58 +8,33 @@ namespace PicoDotNet.Builder.Runtime;
 
 public static class Program
 {
-    private static string Nasm, Gcc, Ld, Objdump, LimineDeploy, Qemu;
-
     private const string KernelArgs =
         "-nostdlib -ffreestanding -Wall -Wextra -fno-exceptions -Wno-write-strings -Wno-unused-parameter -Wno-unused-variable";
 
     public static void Main(string[] _)
     {
-        Console.WriteLine(Environment.CurrentDirectory);
         Directory.SetCurrentDirectory("../PicoDotNet.Runtime.C");
 
-        if (RuntimeInformation.IsOSPlatform(OSPlatform.Windows))
-        {
-            Nasm = Path.Combine("Build", "nasm.exe");
-            Gcc = Path.Combine("Build", "i686-elf-tools-windows", "bin", "i686-elf-gcc.exe");
-            Ld = Path.Combine("Build", "i686-elf-tools-windows", "bin", "i686-elf-ld.exe");
-            Objdump = Path.Combine("Build", "i686-elf-tools-windows", "bin", "i686-elf-objdump.exe");
-            LimineDeploy = Path.Combine("Build", "limine-deploy.exe");
-            Qemu = Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.ProgramFiles), "qemu",
-                "qemu-system-i386");
+        BuildPaths.LoadConfiguration();
 
-            if (!File.Exists(Gcc))
-                Console.WriteLine("Please download i686-elf-tools-windows and put in inside the Build directory of the builder!\nYou can download it here: https://github.com/alessandromrc/i686-elf-tools/releases");
-        }
-        else
-        {
-            Nasm = "nasm";
-            Gcc = "i686-elf-gcc";
-            Ld = "i686-elf-ld";
-            Objdump = "i686-elf-objdump";
-            LimineDeploy = "limine-deploy";
-            Qemu = "qemu-system-i386";
-        }
-
-        if (Directory.Exists("Bin"))
-            Directory.Delete("Bin", true);
+        if (Directory.Exists("Bin")) { Directory.Delete("Bin", true); }
         Directory.CreateDirectory("Bin/Obj");
 
-        Process.Start(Nasm, "-felf32 Source/Start.asm -o Bin/Start.o").WaitForExit();
+        Process.Start(BuildPaths.Nasm, "-felf32 Source/Start.asm -o Bin/Start.o").WaitForExit();
 
         foreach (var file in Directory.GetFiles("Source", "*.c", SearchOption.AllDirectories))
-            Process.Start(Gcc,
+            Process.Start(BuildPaths.Gcc,
                     $"-Iinclude -I../libk/include -c \"{file}\" -o \"Bin/Obj/{Path.GetFileNameWithoutExtension(file) + ".o"}\" {KernelArgs}")
                 .WaitForExit();
 
-        Process.Start(Ld,
+        Process.Start(BuildPaths.Ld,
             Directory.GetFiles("Bin/Obj", "*.o", SearchOption.AllDirectories).Aggregate(
                 "-T \"Build/linker.ld\" -o \"Bin/Kernel.elf\" \"Bin/Start.o\"",
                 (current, file) => current + $" \"{file}\"")).WaitForExit();
 
         var p = Process.Start(new ProcessStartInfo
         {
-            FileName = Objdump,
+            FileName = BuildPaths.Objdump,
             Arguments = "-t -dwarf -Mintel \"Bin/Kernel.elf\"",
             UseShellExecute = false,
             CreateNoWindow = false,
@@ -86,8 +61,8 @@ public static class Program
         iso.SetBootImage(cd, BootDeviceEmulation.NoEmulation, 0);
         iso.Build("Bin/PicoDotNet.iso");
 
-        Process.Start(LimineDeploy, "--force-mbr Bin/PicoDotNet.iso").WaitForExit();
-        Process.Start(Qemu,
+        Process.Start(BuildPaths.LimineDeploy, "--force-mbr Bin/PicoDotNet.iso").WaitForExit();
+        Process.Start(BuildPaths.Qemu,
                 "-usb -m 2G -cdrom Bin/PicoDotNet.iso -serial stdio -device AC97 -rtc base=localtime -no-reboot -no-shutdown")
             .WaitForExit();
     }
